Put LSP hover functionality in its own file
