+++ util/lsp/src/lib.rs
@@ -1,6 +1,7 @@
 mod capabilities;
 mod conversion;
 mod diagnostics;
+mod hover;
 mod server;
 
 pub use server::*;
