+++ util/lsp/src/server.rs
@@ -4,7 +4,7 @@ use tower_lsp::jsonrpc::Result;
 use tower_lsp::{jsonrpc, lsp_types::*, LanguageServer};
 
 use printer::{PrintCfg, PrintToString};
-use query::{Binder, Ctx, Database, DatabaseView, File, HoverInfo, Xfunc};
+use query::{Database, File, Xfunc};
 
 use super::capabilities::*;
 use super::conversion::*;
@@ -62,15 +62,7 @@ impl LanguageServer for Server {
     }
 
     async fn hover(&self, params: HoverParams) -> jsonrpc::Result<Option<Hover>> {
-        let pos_params = params.text_document_position_params;
-        let text_document = pos_params.text_document;
-        let pos = pos_params.position;
-        let db = self.database.read().await;
-        let index = db.get(text_document.uri.as_str()).unwrap();
-        let info =
-            index.location_to_index(pos.from_lsp()).and_then(|idx| index.hoverinfo_at_index(idx));
-        let res = info.map(|info| info_to_hover(&index, info));
-        Ok(res)
+        super::hover::hover(self, params).await
     }
 
     async fn code_action(
@@ -148,47 +140,4 @@ impl Server {
     async fn send_diagnostics(&self, url: Url, diags: Vec<Diagnostic>) {
         self.client.publish_diagnostics(url, diags, None).await;
     }
-}
-
-fn info_to_hover(index: &DatabaseView, info: HoverInfo) -> Hover {
-    let range = info.span.and_then(|span| index.span_to_locations(span)).map(ToLsp::to_lsp);
-
-    let contents = match info.ctx {
-        Some(ctx) => {
-            let mut value = String::new();
-            goal_to_markdown(&info.typ, &mut value);
-            value.push_str("\n\n");
-            ctx_to_markdown(&ctx, &mut value);
-            HoverContents::Markup(MarkupContent { kind: MarkupKind::Markdown, value })
-        }
-        None => HoverContents::Scalar(MarkedString::LanguageString(LanguageString {
-            language: "pol".to_owned(),
-            value: info.typ,
-        })),
-    };
-
-    Hover { contents, range }
-}
-
-fn ctx_to_markdown(ctx: &Ctx, value: &mut String) {
-    value.push_str("**Context**\n\n");
-    value.push_str("| | |\n");
-    value.push_str("|-|-|\n");
-    for Binder { name, typ } in ctx.bound.iter().rev().flatten() {
-        if name == "_" {
-            continue;
-        }
-        value.push_str("| ");
-        value.push_str(name);
-        value.push_str(" | `");
-        value.push_str(typ);
-        value.push_str("` |\n");
-    }
-}
-
-fn goal_to_markdown(goal_type: &str, value: &mut String) {
-    value.push_str("**Goal**\n\n");
-    value.push_str("```\n");
-    value.push_str(goal_type);
-    value.push_str("\n```\n");
 }
